// src/configs/global.enum.ts
export enum verificationActionType {
  SEND_VERIFICATION_CODE_CALLER = 'DO-SEND-VERIFICATION-CODE ',
  SEND_VERIFICATION_CODE_DEFAULT = 'SEND-VERIFICATION-CODE-NOT-STARTED',
  SEND_VERIFICATION_CODE_STARTED = 'SEND-VERIFICATION-CODE-STARTED',
  SEND_VERIFICATION_CODE_FAILED = 'SEND-VERIFICATION-CODE-FAILED',
  SEND_VERIFICATION_CODE_SUCCESS = 'SEND-VERIFICATION-CODE-SUCCESS',
}
// ...
export enum RessetPasswordActionType {
  RESSET_PASSWORD_CALLER = 'DO-RESSET-PASSWORD ',
  RESSET_PASSWORD_DEFAULT = 'RESSET-PASSWORD-NOT-STARTED',
  RESSET_PASSWORD_STARTED = 'RESSET-PASSWORD-STARTED',
  RESSET_PASSWORD_FAILED = 'RESSET-PASSWORD-FAILED',
  RESSET_PASSWORD_SUCCESS = 'RESSET-PASSWORD-SUCCESS',
}
// ...
export enum verifyingCodeAction {
  VERIFYING_CODE_CALLER = 'DO-VERIFY-CODE',
  VERIFYING_CODE_DEFAULT = 'DO-VERIFY-CODE',
  VERIFYING_CODE_STARTED = 'VERIFY-CODE-STARTED',
  VERIFYING_CODE_FAILED = 'VERIFY-CODE-FAILED',
  VERIFYING_CODE_SUCCESS = 'VERIFY-CODE-SUCCESS',
}

export enum loginActionType {
  LOGIN_CALLER = 'DO-LOGIN',
  LOGIN_STARTED = 'LOGIN-STARTED',
  LOGIN_SUCCESS = 'LOGIN-SUCCESS',
  LOGIN_FAILED = 'LOGIN-FAILED',
  LOGIN_DEFAULT = 'LOGIN-NOT-STARTED',
}

export enum signUpActionType {
  SIGNUP_CALLER = 'DO-SIGNUP',
  SIGNUP_STARTED = 'SIGNUP-STARTED',
  SIGNUP_SUCCESS = 'SIGNUP-SUCCESS',
  SIGNUP_FAILED = 'SIGNUP-FAILED',
  SIGNUP_DEFAULT = 'SIGNUP-NOT-STARTED',
}

export enum currentLocation {
  GET_CUURENT_LOCATION_CALLER = 'DO-GET-CURRENT-LOCATION',
  GET_CUURENT_LOCATION_STARTED = 'GET-CURRENT-LOCATION-STARTED',
  GET_CUURENT_LOCATION_SUCCESS = 'GET-CURRENT-LOCATION-SUCCESS',
  GET_CUURENT_LOCATION_FAILED = 'GET-CURRENT-LOCATION-FAILED',
  GET_CUURENT_LOCATION_DEFAULT = 'GET-CURRENT-LOCATION-NOT-STARTED',
}

export enum inputActionType {
  //SAGA CALLERS
  SET_LOGIN_PHONE_NUMBER_CALLER = 'SET-LOGIN-PHONE-NUMBER',
  SET_LOGIN_PASSWORD_CALLER = 'SET-LOGIN-PASSWORD',
  SET_FIRST_NAME_CALLER = 'SET-FIRST-NAME',
  SET_LAST_NAME_CALLER = 'SET-LAST-NAME',
  SET_SIGNUP_PHONE_NUMBER_CALLER = 'SET-SIGNUP-PHONE-NUMBER',
  SET_SIGNUP_PASSWORD_CALLER = 'SET-SIGNUP-PASSWORD',
  SET_GENDER_CALLER = 'SET-GENDER',
  SET_RESSET_PASSWORD_PHONE_NUMBER_CALLER = 'SET-RESSET-PASSWORD-PHONE-NUMBER',
  SET_VERIFICATION_PIN_CALLER = 'SET-VERIFICATION-PIN',
  SET_TIMER_ELAPSE_CALLER = 'SET-TIMER-ELAPSE',
  CHANGE_COUNTDOWN_COMPONENT_ID_CALLER = 'CHANGE-COUTDOWN-COMPOENT-ID',
  SET_NEW_PASSWORD_CALLER = 'SET-NEW-PASSWORD',
  SET_CONFIRM_NEW_PASSWORD_CALLER = 'SET-CONFIRM-NEW-PASSWORD',
  SET_PARCEL_LOCATION_CALLER = 'DO-SET-PARCEL-LOCATION',
  SET_PARCEL_DESITINATION_CALLER = 'DO-SET-PARCEL-DESTINATION',
  SET_LOCATION_ACTIVE_CALLER = 'DO-SET-LOCATION-ACTIVE',
  SET_VARIABLES_CALLER = 'DO-SET-VARIABLES',
  SET_RESET_INPUTS_CALLER = 'DO-RESET-INPUTS',
  SET_DURATION_AND_DISTANCE_CALLER = 'DO-SET-DURATION-AND-DISTANCE',
  SET_ENABLE_ADD_BUTTON_CALLER = 'DO-ENABLE-ADD-BUTTON',
  SET_ADD_VALID_CARD_CALLER = 'DO-ADD-VALID-CARD',
  SET_REMOVE_VALID_CARD_CALLER = 'DO-REMOVE-VALID-CARD',
  SET_CHANGE_PAYMENT_METHOD_CALLER = 'DO-CHANGE-PAYMENT-METHOD',
  SET_TRACK_ACTIVE_DELIVERY_CALLER = 'DO-SET-DELIVERY-ON-PROGRESS',
  SET_USER_CURRENT_LOCATION_CALLER = 'DO-SET-CURRENT-LOCATION',
  SET_LOCATION_CALLER = 'DO-SET-FETCHED-LOCATION',
  SET_PROFILE_PICTURE_CALLER = 'DO-SET-USER-PROFILE-PICTURE',
  SET_CARD_EMAIL_CALLER = 'DO_SET_CARD_EMAIL_CALLER',

  //ACTIONS
  SET_LOGIN_PHONE_NUMBER = 'SET-LOGIN-PHONE-NUMBER',
  SET_LOGIN_PASSWORD = 'SET-LOGIN-PASSWORD',
  SET_FIRST_NAME = 'SET-FIRST-NAME',
  SET_LAST_NAME = 'SET-LAST-NAME',
  SET_SIGNUP_PHONE_NUMBER = 'SET-SIGNUP-PHONE-NUMBER',
  SET_SIGNUP_PASSWORD = 'SET-SIGNUP-PASSWORD',
  SET_GENDER = 'SET-GENDER',
  SET_RESSET_PASSWORD_PHONE_NUMBER = 'SET-RESSET-PASSWORD-PHONE-NUMBER',
  SET_VERIFICATION_PIN = 'SET-VERIFICATION-PIN',
  SET_TIMER_ELAPSE = 'SET-TIMER-ELAPSE',
  CHANGE_COUNTDOWN_COMPONENT_ID = 'CHANGE-COUTDOWN-COMPOENT-ID',
  SET_NEW_PASSWORD = 'SET-NEW-PASSWORD',
  SET_CONFIRM_NEW_PASSWORD = 'SET-CONFIRM-NEW-PASSWORD',
  SET_PARCEL_LOCATION = 'SET-PARCEL-LOCATION',
  SET_PARCEL_DESITINATION = 'SET-PARCEL-DESTINATION',
  SET_LOCATION_ACTIVE = 'SET-LOCATION-ACTIVE',
  SET_VARIABLES = 'SET-VARIABLES',
  SET_RESET_INPUTS = 'RESET-INPUTS',
  SET_DURATION_AND_DISTANCE = 'SET-DURATION-AND-DISTANCE',
  SET_ENABLE_ADD_BUTTON = 'ENABLE-ADD-BUTTON',
  SET_ADD_VALID_CARD = 'ADD-VALID-CARD',
  SET_REMOVE_VALID_CARD = 'REMOVE-VALID-CARD',
  SET_CHANGE_PAYMENT_METHOD = 'CHANGE-PAYMENT-METHOD',
  SET_TRACK_ACTIVE_DELIVERY = 'SET-DELIVERY-ON-PROGRESS',
  SET_USER_CURRENT_LOCATION = 'SET-CURRENT-LOCATION',
  SET_PROFILE_PICTURE = 'SET-USER-PROFILE-PICTURE',
  SET_CARD_EMAIL = 'SET_CARD_EMAIL',
}